+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Utilities: Add media queries to VisuallyHidden utility.
+
 ## [12.7.0] - 2022-09-05 👽
 
 Feature: